Make accessor overrides call #super if available
Some classes may have overridden the default Ruby accessors (setters
and getters) in order to transform the data, perform caching,
perform additional processing on set/get, etc. When using Rails,
classes that inherit from ActiveRecord should ideally call the
superclass's accessor to ensure that things run smoothly. This
fix ensures that those custom accessors are called, if available.

The accessors of attributes that use SymmetricEncryption are overridden
to allow transparent encryption and decryption. I.e., When setting an
attribute in plaintext, it is automatically encrypted in the
background; and when retrieving a value, it is transparently decrypted
as needed.

By setting the instance variables directly, these custom accessors were
being bypassed. This can lead to unexpected consequences. The type of
bugs that come from this can be particularly sneaky and difficult to
track down.

This commit fixes the issue where custom accessors added in an
overridden class, or provided by a framework, were being skipped.