Bring Reader memory usage to a constant
These changes shave off memory usage by two more filesizes, making it
effectively constant regardless of the size of the encrypted file.

There are two main changes here:

* We avoid accumulating multiple chunks in @read_buffer, as that seems
  to cause some problems with proper memory deallocation. This shaves
  off one filesize of memory usage.

* We make #read retrieve number of bytes that are needed from the
  underlying IO, which reduces memory allocations because it removes the
  need to slicing the string. This shaves off another filesize of memory
  usage.

For this we needed to refactor the #read method significantly (and some
other significantly as well), this refactoring was based on the
implementation of Down::ChunkedIO.

https://github.com/janko-m/down/blob/master/lib/down/chunked_io.rb

This refactoring also reduces duplication in decrypting chunks, and
fixes #seek calling nonexisting #eof method when IO::SEEK_END is used.