Support output buffer in Reader#read and drastically reduce Reader & Writer memory usage (#98)
* Accept output buffer in Reader#read

The IO#read method can accept an optional second argument – the output
buffer. When the output buffer string is provided, each new chunk gets
read into it (replacing any previous content), reducing string
allocations.

This commits extends SymmetricEncryption::Reader#read with the ability
to accept an output buffer. In addition to allowing one to reduce string
allocations, it also enables the SymmetricEncryption::Reader to be the
first argument to IO.copy_stream (which assumes #read accepts an output
buffer).

* Simplify Writer.encrypt and Reader.decrypt

We can simplify the logic of these two methods a lot by using
IO.copy_stream. There are few things to note here:

* Since Reader#read now accepts an output buffer, we can use the Reader
  object as the first argument in IO.copy_stream. This reduces string
  allocations both in Writer.encrypt and Reader.decrypt.

* IO.copy_stream accepts an IO object or a path for both the source and
  destination arguments, and under the hood the files will automatically
  be opened in binary mode, so don't need do this logic manually
  anymore.

* IO.copy_stream returns number of bytes written, so it will be
  implicitly returned from this method.

* There is no need to close the source/target anymore, because if they
  were provided as a path string, IO.copy_stream will take care that
  they're closed after opening. Another benefit of this change is that
  custom source/target IOs won't be closed anymore, allowing the user
  to choose when they want to close their IO.

* Use output buffer internally in Reader

By using an output buffer when reading chunks of the source file we
reduce memory allocations, because instead of Ruby allocating a string
for each new chunk, the output buffer string will be reused for each new
chunk.

* Deallocate decrypted chunks in Reader

This reduces the memory usage by one filesize of the encrypted file.

* Reduce memory usage when output buffer is used

It seems there is a memory bug with String#replace were it's not
properly deallocating the replaced string, and using String#clear
appears to fix this. See this bug report for more details:

https://bugs.ruby-lang.org/issues/14745

This reduces memory usage by another filesize of the encrypted file.

* Bring Reader memory usage to a constant

These changes shave off memory usage by two more filesizes, making it
effectively constant regardless of the size of the encrypted file.

There are two main changes here:

* We avoid accumulating multiple chunks in @read_buffer, as that seems
  to cause some problems with proper memory deallocation. This shaves
  off one filesize of memory usage.

* We make #read retrieve number of bytes that are needed from the
  underlying IO, which reduces memory allocations because it removes the
  need to slicing the string. This shaves off another filesize of memory
  usage.

For this we needed to refactor the #read method significantly (and some
other significantly as well), this refactoring was based on the
implementation of Down::ChunkedIO.

https://github.com/janko-m/down/blob/master/lib/down/chunked_io.rb

This refactoring also reduces duplication in decrypting chunks, and
fixes #seek calling nonexisting #eof method when IO::SEEK_END is used.

* Halve memory usage of Writer

String#clear in MRI immediately releases memory allocates from
malloc(3), so it can make a lot of difference in keeping memory usage
low. See this post by Eric Hodel from the ruby-talk mailing list:

https://rubytalk.org/t/psa-string-memory-use-reduction-techniques/74477

This commit clears the encrypted chunks of data after they written to
the underlying IO object in `SymmetricEncryption::Writer#write`. This
halves the memory usage of `Writer.encrypt`; previously encrypting 50MB
file would use about 100MB of memory, and with this change it uses only
50MB.

* Use a buffer for cipher updates

* Use binary output buffers

OpenSSL::Cipher#update seems to have some problems when binary data is
encrypted if buffer is not binary (it was uncovered in tests that use
compression).