Clarify some process in configuration.md
The order and details of a few steps were misleading. Tried to correct. For future work, someone should clarify the difference between performing the initial configuration on a localhost and then moving to a production server and when configuration is done directly on the production host. The second edit attempts to help in that regard.